Pushing this before scraping and attempting solution #2
Using WorldPortalRenderer in iChun leads to crashing from calling TileEntityRendererDispatcher.instance.drawBatch() whenever I try to update the portal. I think because this is an entity and not a tile entity, the draw order gets all fucky. I'm going to instead create a second RenderGlobal and "Portal View Entity" and do that instead