New Render Attempt #2
Tried using the code from the mirror in CFBlocks, but changed the viewing entity to a new "Portal View Entity" that's created when a portal is created. it at least renders the world, but only if the player is in tracking distance of the portal view entity, so it's not actually feasible for long term use.

I'm instead going to save this as a branch and go back to the old way, see if I can figure out where I'm wrong in that render process. iChun has at least figured out how to get far away chunks to render